python: adding the --list-tests and --mvn-mode flags.
The --list-tests flag, is printing a list of all the TestCaseClass.TestMethod
which are found in the file. This way each test can be called separately.

The --mvn-mode is used to assume that the call is done via the PlanBuilder class
(the usuall mvn way) and this is causing the utils.py and subprocess_testcase.py to be copied
to the remote node, thus the import path is a bit different.